+++ client/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Ensure autofill works for login page
+
 ## [0.3.3] - 2025-04-29
 ### Changed
 - Updated dependencies
@@ -104,6 +108,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
+[Unreleased]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.3...vHEAD
 [0.3.3]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.2...v0.3.3
 [0.3.2]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.1...v0.3.2
 [0.3.1]: https://github.com/Skycoder42/systemd_status/compare/client%2Fv0.3.0...v0.3.1